requirements.txt
ipyleaflet>=0.14.0
ipykernel>=6.29.0
ipywidgets>=7.7.1
matplotlib>=3.7.0
numpy>=1.24.2
opencv-python>=4.1.2
pandas>=2.3.3
psycopg2-binary>=2.9.3
pymysql>=1.0.2
python-dotenv>=0.20.0
requests>=2.27.1
scikit-learn>=1.5.0,<2.0.0
seaborn>=0.12.2
sqlalchemy>=2.0.38
sympy>=1.10.1
xgboost>=2.1.0,<3.0.0
